Hacked version of skipIntro feature.
Allows the user to skip most intro animations by specifying "skipIntro=true" in
their settings.json file.

This speeds up the development workflow, since you don't have to sit through a
few seconds of animation when the app boots up.

However, it's a total hack in this state, because it requires duplicating a lot
of logic between the skipIntro and intro codepaths. I need to find a way of
designing this that's less hacky.